server/requirements.txt
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.1
filelock==3.13.4
fsspec==2024.3.1
gradio_client==0.16.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
packaging==24.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
websockets==11.0.3